Restructure view models in separate files in same folder
